services/feed_service/requirements.txt
amqp==5.2.0
async-timeout==4.0.3
billiard==4.2.1
blinker==1.8.2
boto3==1.35.30
botocore==1.35.30
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ffmpeg-python==0.2.0
Flask==3.0.3
Flask-Cors==5.0.0
future==1.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.4.2
loguru==0.7.2
MarkupSafe==2.1.5
pika==1.3.2
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.1.0
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
tzdata==2024.2
urllib3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
redis
